neucipher-frontend/src/app/login-doctor/login-doctor.component.ts
import { Component } from '@angular/core';
import { AuthService } from '../auth.service';

@Component({
  selector: 'app-login-doctor',
  templateUrl: './login-doctor.component.html',
  styleUrls: ['./login-doctor.component.scss']
})
export class LoginDoctorComponent {
  email: string = '';
  password: string = '';

  constructor(private authService: AuthService) { }

  onLogin() {

    let userData = {
      email : this.email,
      password : this.password
    }
    this.authService.logindoctor(userData);
  }
}
